pyproject.toml
[tool.poetry]
name = "medium-01-celery-example"
version = "0.1.0"
description = ""
authors = ["Artur Rakhmatulin <artur.rakhmatulin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-results = "^2.5.1"
djangorestframework = "^3.15.1"
gunicorn = "^22.0.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
